Split ProjectWindow.py into status_bar, output_tabs, guided_pane, and project_window_menu modules
ProjectWindow.py had grown to ~1826 lines, bundling several largely
independent widget classes (status bar, output-tab viewer, the entire
guided-mode input builder) and menu-bar construction alongside the
ProjectWindow class itself. Move each into its own module with no
behavior change; ProjectWindow keeps only its own class body and now
delegates menu-bar setup to project_window_menu.setup_project_window_menubar.
